Introduced map.impl.event package
To reduce the size of the map.impl package, a subpackage is created where the event
logic is copied into.
