+++ hazelcast/src/main/java/com/hazelcast/map/impl/event/EntryEventData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map.impl;
+package com.hazelcast.map.impl.event;
 
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
