+++ hazelcast/src/main/java/com/hazelcast/map/impl/event/MapEventData.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.hazelcast.map.impl;
+package com.hazelcast.map.impl.event;
 
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.ObjectDataInput;
@@ -23,43 +23,41 @@ import com.hazelcast.nio.ObjectDataOutput;
 import java.io.IOException;
 
 /**
- * Contains the data related to a map partition event
+ * Map wide event's data.
  */
-public class MapPartitionEventData extends AbstractEventData {
+public class MapEventData extends AbstractEventData {
 
-    private int partitionId;
+    protected int numberOfEntries;
 
-    public MapPartitionEventData() {
+    public MapEventData() {
     }
 
-    public MapPartitionEventData(String source, String mapName, Address caller, int partitionId) {
-        super(source, mapName, caller, -1);
-        this.partitionId = partitionId;
+    public MapEventData(String source, String mapName, Address caller, int eventType, int numberOfEntries) {
+        super(source, mapName, caller, eventType);
+        this.numberOfEntries = numberOfEntries;
     }
 
-    public int getPartitionId() {
-        return partitionId;
+    public int getNumberOfEntries() {
+        return numberOfEntries;
     }
 
     @Override
-    public void writeData(ObjectDataOutput out)
-            throws IOException {
+    public void writeData(ObjectDataOutput out) throws IOException {
         super.writeData(out);
-        out.writeInt(partitionId);
+        out.writeInt(numberOfEntries);
     }
 
     @Override
-    public void readData(ObjectDataInput in)
-            throws IOException {
+    public void readData(ObjectDataInput in) throws IOException {
         super.readData(in);
-        partitionId = in.readInt();
+        numberOfEntries = in.readInt();
     }
 
     @Override
     public String toString() {
-        return "MapPartitionEventData{"
+        return "MapEventData{"
                 + super.toString()
-                + ", partitionId=" + partitionId
+                + ", numberOfEntries=" + numberOfEntries
                 + '}';
     }
 }
