+++ hazelcast/src/main/java/com/hazelcast/map/impl/querycache/event/QueryCacheEventData.java
@@ -16,7 +16,7 @@
 
 package com.hazelcast.map.impl.querycache.event;
 
-import com.hazelcast.map.impl.EventData;
+import com.hazelcast.map.impl.event.EventData;
 import com.hazelcast.map.impl.querycache.event.sequence.Sequenced;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.internal.serialization.SerializationService;
