+++ hazelcast/src/main/java/com/hazelcast/multimap/impl/MultiMapEventsPublisher.java
@@ -17,8 +17,8 @@
 package com.hazelcast.multimap.impl;
 
 import com.hazelcast.core.EntryEventType;
-import com.hazelcast.map.impl.EntryEventData;
-import com.hazelcast.map.impl.MapEventData;
+import com.hazelcast.map.impl.event.EntryEventData;
+import com.hazelcast.map.impl.event.MapEventData;
 import com.hazelcast.nio.Address;
 import com.hazelcast.nio.serialization.Data;
 import com.hazelcast.spi.EventRegistration;
