+++ hazelcast/src/test/java/com/hazelcast/map/ListenerTest.java
@@ -26,7 +26,7 @@ import com.hazelcast.core.EntryListener;
 import com.hazelcast.core.HazelcastInstance;
 import com.hazelcast.core.IMap;
 import com.hazelcast.core.MapEvent;
-import com.hazelcast.map.impl.MapPartitionEventData;
+import com.hazelcast.map.impl.event.MapPartitionEventData;
 import com.hazelcast.map.impl.MapService;
 import com.hazelcast.map.listener.EntryUpdatedListener;
 import com.hazelcast.map.listener.MapPartitionLostListener;
